make rmdotfiles optional

+++ normalize-plex-files/utils.py
@@ -2,6 +2,8 @@ import glob
 import os
 import sys
 from pathlib import Path
+from types import SimpleNamespace
+
 
 def resolutionstring(height: int, width: int) -> str:
     """returns a string of the form " [{width}x{height}]".
@@ -51,19 +53,21 @@ def basedir(configuredbase: str, currentfile: str, depth: int = 1) -> str:
 
 
 
-def movemedia(old_file:str, new_file:str, armed:bool) -> None:
+def movemedia(old_file:str, new_file:str, config:SimpleNamespace) -> None:
     """Moves all files with basename {old_file} and arbitrary extensions
-    to {new_file} retaining the extensions. {old_file} must not contain an
-    extension.
+    to {new_file} retaining the extensions when {config.armed} is true.
+    {old_file} must not contain an filename-extension.
     Does not overwrite existing files. Removes the old directory if empty after
-    operation, forcefully removes dot-files from the old directory to do so."""
+    operation.
+    Forcefully removes dot-files from the old directory to allow for directory
+    removal if {config.rmdotfiles} is true."""
     # old_file and new_file are basenames without file extensions.
     # The move concerns all files with these basenames, regardless of extension.
 
     # check if file should be actually moved
     if old_file != new_file:
         # yes!
-        if not armed:
+        if not config.armed:
             print(f"would move: {old_file} --> {new_file}")
         else:
             # seek all extensions of old_file (e.g. .m4v, .srt, ...)
@@ -83,16 +87,17 @@ def movemedia(old_file:str, new_file:str, armed:bool) -> None:
                     else:
                         # try to remove old dir
                         # (will fail, if this is not the last movie file)
-                        # prereq: remove dot files in dir
+                        # prereq: remove dot files in dir, if config.rmdotfiles
                         old_base_dir = os.path.dirname(old_file)
-                        dotfiles = glob.glob(os.path.join(
-                            glob.escape(old_base_dir), '.??*'))
-                        for dotfile in dotfiles:
-                            print(f"removing {dotfile}")
-                            try:
-                                os.unlink(dotfile)
-                            except Exception as e:
-                                print(e, file=sys.stderr)
+                        if config.rmdotfiles:
+                            dotfiles = glob.glob(os.path.join(
+                                glob.escape(old_base_dir), '.??*'))
+                            for dotfile in dotfiles:
+                                print(f"removing {dotfile}")
+                                try:
+                                    os.unlink(dotfile)
+                                except Exception as e:
+                                    print(e, file=sys.stderr)
                         # actual rmdir:
                         # ignore if not empty (this only indicates that this file
                         # was not the last movie file. Removal will be successful
